split some implementation into different files for reusability
